Add base test that classic and dutch can inherit from
